+++ src/main/java/com/gotcha/domain/block/exception/BlockException.java
@@ -1,0 +1,31 @@
+package com.gotcha.domain.block.exception;
+
+import com.gotcha._global.exception.BusinessException;
+import com.gotcha._global.exception.ErrorCode;
+
+public class BlockException extends BusinessException {
+
+    private BlockException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    private BlockException(ErrorCode errorCode, String additionalInfo) {
+        super(errorCode, additionalInfo);
+    }
+
+    public static BlockException cannotBlockSelf() {
+        return new BlockException(BlockErrorCode.CANNOT_BLOCK_SELF);
+    }
+
+    public static BlockException alreadyBlocked() {
+        return new BlockException(BlockErrorCode.ALREADY_BLOCKED);
+    }
+
+    public static BlockException notFound() {
+        return new BlockException(BlockErrorCode.BLOCK_NOT_FOUND);
+    }
+
+    public static BlockException notFound(Long blockedUserId) {
+        return new BlockException(BlockErrorCode.BLOCK_NOT_FOUND, "blockedUserId: " + blockedUserId);
+    }
+}